+++ nes-execution-engine/src/Experimental/Interpreter/Operators/Scan.cpp
@@ -17,8 +17,7 @@
 #include <Nautilus/Interface/Record.hpp>
 namespace NES::Nautilus {
 
-Scan::Scan(const Runtime::MemoryLayouts::MemoryLayoutPtr memoryLayout,
-           std::vector<Record::RecordFieldIdentifier> projections)
+Scan::Scan(const Runtime::MemoryLayouts::MemoryLayoutPtr memoryLayout, std::vector<Record::RecordFieldIdentifier> projections)
     : memoryLayout(memoryLayout), projections(projections) {}
 
 void Scan::open(RuntimeExecutionContext& ctx, RecordBuffer& recordBuffer) const {
@@ -27,7 +26,7 @@ void Scan::open(RuntimeExecutionContext& ctx, RecordBuffer& recordBuffer) const 
     // iterate over records in buffer
     auto numberOfRecords = recordBuffer.getNumRecords();
     auto bufferAddress = recordBuffer.getBuffer();
-    for (Value<UInt64> i = 0ul; i < numberOfRecords; i = i + 1ul) {
+    for (Value<UInt64> i = (uint64_t) 0; i < numberOfRecords; i = i + (uint64_t) 1) {
         auto record = recordBuffer.read(memoryLayout, projections, bufferAddress, i);
         child->execute(ctx, record);
     }
